alfunduqapi/requirements.txt
Django==2.1.2
pytz==2018.5
mysqlclient==1.3.13
djangorestframework==3.8.2
django-rest-auth==0.9.3
django-allauth
django-cors-headers==2.4.0
django-rest-swagger==2.2.0